feat: export functions to find aws vars separate
 * Can now not from within a serverless template.
